Add useToolBar hook
